pyproject.toml
[project]
name = "du-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "httpx>=0.28.1",
    "pyarrow>=19.0.1",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "scrapy>=2.12.0",
    "selenium>=4.30.0",
    "webdriver-manager>=4.0.2",
    "xmltodict>=0.14.2",
]

[dependency-groups]
dev = [
    "apache-airflow>=2.10.5",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
]
